pyproject.toml
[project]
name = "mcp-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["anyio>=4.5", "click>=8.1.0", "httpx>=0.27", "mcp", "starlette", "uvicorn", "fastapi", "langchain-mcp-adapters>=0.0.8", "langchain-core", "langchain-google-genai", "langchain-community", "langchain", "python-dotenv"]

[project.scripts]
mcp_math = "mcp_math.server:main"
mcp_weather = "mcp_weather.server:main"

[tool.uv]
package = true

[tool.setuptools.packages.find]
include = ["mcp_math", "mcp_weather"]
